refactor command execution logic to include aliases
the onMessage method now grabs an instance of a command based on the raw command name, or any one of it's aliases. the aliases are assigned in the classes as a property, and it's value is an array of strings (i.e. aliases = ['!sp', '!stp']). The execution then happens based on the command instance grabbed, dynamically.